Add SP3 subtest support to run-benchmark https://bugs.webkit.org/show_bug.cgi?id=288652 rdar://145692860
Reviewed by Dewei Zhu.

Add SP3 subtests to the SP3 plan file.
Slightly adjust how plan files are used to construct a url.
(SP3 subtests aren't separated by '&'.)
(In SP3, suites are runnable rather than tests within suites.)
Adjust other benchmark plan files to compensate for this.
Refactor _construct_subtest_url to parent class.

* Tools/Scripts/webkitpy/benchmark_runner/data/plans/jetstream2.1.plan:
* Tools/Scripts/webkitpy/benchmark_runner/data/plans/jetstream2.2.plan:
* Tools/Scripts/webkitpy/benchmark_runner/data/plans/speedometer3.plan:
* Tools/Scripts/webkitpy/benchmark_runner/webserver_benchmark_runner.py:
(WebServerBenchmarkRunner._construct_subtest_url):

Canonical link: https://commits.webkit.org/291516@main
