+++ Tools/Scripts/webkitpy/benchmark_runner/data/plans/jetstream2.1.plan
@@ -518,7 +518,7 @@
             {"path": "worker/segmentation.js", "mode": "100644", "type": "blob"}
         ]
     },
-    "subtest_url_format": "test=${TEST}",
+    "subtest_url_format": "&test=${TEST}",
     "subtests": {
         "": [
             "Air",
