+++ Tools/Scripts/webkitpy/benchmark_runner/data/plans/jetstream2.2.plan
@@ -518,7 +518,7 @@
         ],
         "truncated": false
     },
-    "subtest_url_format": "test=${TEST}",
+    "subtest_url_format": "&test=${TEST}",
     "subtests": {
         "": [
             "Air",
