+++ Tools/Scripts/webkitpy/benchmark_runner/data/plans/motionmark1.2.1.plan
@@ -7,7 +7,7 @@
     "webdriver_benchmark_patch": "data/patches/webdriver/MotionMark.patch",
     "signpost_patch": "data/patches/signposts/MotionMark.patch",
     "entry_point": "index.html",
-    "subtest_entry_point": "developer.html",
+    "subtest_entry_point": "developer.html?",
     "config": {
         "orientation": "landscape"
     },
